backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
peewee>=3.17.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
